superadmin added

+++ src/Pages/Dashboard/Admins/Admins.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useContext, useState } from 'react';
 import { FaPlus } from 'react-icons/fa';
 import Btn from '../../../components/Btn';
 import { TextInput } from 'flowbite-react';
@@ -9,6 +9,8 @@ import ViewAdminModal from './Components/ViewAdminModal';
 import { useQuery } from '@tanstack/react-query';
 import { toast } from 'react-hot-toast';
 import { useUtils } from '../../../contexts/UtilsProvider';
+import { AuthContext } from '../../../contexts/AuthProvider/AuthProvider';
+import useRole from '../../../hooks/useRole/useRole';
 
 const Admins = () => {
     const { data: admins = [], refetch } = useQuery({
@@ -16,6 +18,10 @@ const Admins = () => {
         queryFn: () => fetch('https://hr-management-server.vercel.app/admins')
             .then(res => res.json())
     })
+
+    const { user, loading } = useContext(AuthContext)
+
+    const [role] = useRole(user.email, loading)
 
     const [addAdminModalVisibility, setAddAdminModalVisibility] = useState(false);
 
@@ -26,21 +32,25 @@ const Admins = () => {
     const [deleteAdmin, setDeleteAdmin] = useState([]);
 
     const handleDelete = () => {
-        fetch(`https://hr-management-server.vercel.app/admins/${deleteAdmin._id}`, {
-            method: "DELETE"
-        })
-            .then((res) => res.json())
-            .then(data => {
-                console.log(data);
-                if (data.deletedCount > 0) {
-                    setDeleteModalVisibility(false);
-                    toast.success(`${deleteAdmin.firstName} deleted`);
-                    refetch();
-                    fetch(`https://hr-management-server.vercel.app/firebase/deleteUser/${deleteAdmin.email}`)
-                        .then((res) => res.json())
-                        .then(data => { })
-                }
-            });
+        if (role) {
+            fetch(`https://hr-management-server.vercel.app/admins/${deleteAdmin._id}`, {
+                method: "DELETE"
+            })
+                .then((res) => res.json())
+                .then(data => {
+                    console.log(data);
+                    if (data.deletedCount > 0) {
+                        setDeleteModalVisibility(false);
+                        toast.success(`${deleteAdmin.firstName} deleted`);
+                        refetch();
+                        fetch(`https://hr-management-server.vercel.app/firebase/deleteUser/${deleteAdmin.email}`)
+                            .then((res) => res.json())
+                            .then(data => { })
+                    }
+                });
+        } else {
+            toast.error('Access denied')
+        }
     };
 
     // Change title
@@ -61,7 +71,7 @@ const Admins = () => {
                     />
                     <Btn color="blue" className="w-fit px-4">Search</Btn>
                 </div>
-                <Btn onClick={() => setAddAdminModalVisibility(!addAdminModalVisibility)} color="blue" className="w-fit ml-auto px-4 flex items-center gap-2 ">Add New <FaPlus /></Btn>
+                <Btn onClick={() => setAddAdminModalVisibility(!addAdminModalVisibility)} color={`${role ? 'blue' : 'gray'}`} className="w-fit ml-auto px-4 flex items-center gap-2" disabled={!role}>Add New <FaPlus /></Btn>
             </div>
             <div className='grid sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-5'>
                 {
